+++ crates/egui/src/containers/combo_box.rs
@@ -77,7 +77,7 @@ impl ComboBox {
         }
     }
 
-    /// Set the width of the button and menu
+    /// Set the outer width of the button and menu.
     pub fn width(mut self, width: f32) -> Self {
         self.width = Some(width);
         self
@@ -261,15 +261,20 @@ fn combo_box_dyn<'c, R>(
             AboveOrBelow::Above
         };
 
+    let margin = ui.spacing().button_padding;
     let button_response = button_frame(ui, button_id, is_popup_open, Sense::click(), |ui| {
         // We don't want to change width when user selects something new
-        let full_minimum_width = wrap_enabled
-            .then(|| ui.available_width() - ui.spacing().item_spacing.x * 2.0)
-            .unwrap_or_else(|| ui.spacing().slider_width);
+        let full_minimum_width = if wrap_enabled {
+            ui.available_width() - ui.spacing().item_spacing.x * 2.0
+        } else {
+            ui.spacing().slider_width - 2.0 * margin.x
+        };
         let icon_size = Vec2::splat(ui.spacing().icon_width);
-        let wrap_width = wrap_enabled
-            .then(|| ui.available_width() - ui.spacing().item_spacing.x - icon_size.x)
-            .unwrap_or(f32::INFINITY);
+        let wrap_width = if wrap_enabled {
+            ui.available_width() - ui.spacing().item_spacing.x - icon_size.x
+        } else {
+            f32::INFINITY
+        };
 
         let galley =
             selected_text.into_galley(ui, Some(wrap_enabled), wrap_width, TextStyle::Button);
@@ -396,16 +401,18 @@ fn paint_default_icon(
     match above_or_below {
         AboveOrBelow::Above => {
             // Upward pointing triangle
-            painter.add(Shape::closed_line(
+            painter.add(Shape::convex_polygon(
                 vec![rect.left_bottom(), rect.right_bottom(), rect.center_top()],
-                visuals.fg_stroke,
+                visuals.fg_stroke.color,
+                Stroke::NONE,
             ));
         }
         AboveOrBelow::Below => {
             // Downward pointing triangle
-            painter.add(Shape::closed_line(
+            painter.add(Shape::convex_polygon(
                 vec![rect.left_top(), rect.right_top(), rect.center_bottom()],
-                visuals.fg_stroke,
+                visuals.fg_stroke.color,
+                Stroke::NONE,
             ));
         }
     }